api/requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.29
Flask-JWT-Extended==4.4.2
python-dotenv==0.20.0